Rework prompting for authentication, to avoid delays
The introduction of Authentication caused the GUI to block while doing a
pre-check that authentication was going to pass (so as to be able to prompt
the user if not). Unfortunately, it ALWAYS blocked the first time, because
it sent a username with a blank password.

This has been reworked to do away with the pre-check. All server access
and blocking again occurs on worker threads.