Fix bug where annotations colud be lost switching between sites
The bug was in forgetting that 'Annotation' was a 'user-field'. So it might be
temporarily written into the 'Tabular' site cache. Upon restarting, it would
correct itself back into the 'local-fields' file. However, if the site was switched
to a different URL while it was only stored in the 'Tabular' file, it would be lost.